renamed metricFindQuery to getVariableQueryValues
