+++ src/datasource.spec.ts
@@ -46,28 +46,28 @@ describe('metricFindQuery', () => {
   const ds = new Datasource(DummyDatasource);
   describe('Random Variable', () => {
     it('Random', async () => {
-      const result = await ds.MetricFindQuery({ query: 'Random(A,B,C,D)', queryType: 'legacy', refId: 'variable' });
+      const result = await ds.getVariableQueryValues({ query: 'Random(A,B,C,D)', queryType: 'legacy', refId: 'variable' });
       expect(['A', 'B', 'C', 'D']).toContain(result[0].text);
     });
   });
 
   describe('Join', () => {
     it('Join', async () => {
-      expect(await ds.MetricFindQuery({ query: 'Join(A,B,C,D)', queryType: 'legacy', refId: 'variable' })).toStrictEqual([{ text: 'ABCD', value: 'ABCD' }]);
+      expect(await ds.getVariableQueryValues({ query: 'Join(A,B,C,D)', queryType: 'legacy', refId: 'variable' })).toStrictEqual([{ text: 'ABCD', value: 'ABCD' }]);
     });
   });
 
   describe('Collection', () => {
     it('Collection', async () => {
-      expect(await ds.MetricFindQuery({ query: 'Collection(A,B,C,D)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
+      expect(await ds.getVariableQueryValues({ query: 'Collection(A,B,C,D)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
     });
   });
 
   describe('CollectionLookup', () => {
     it('CollectionLookup', async () => {
-      expect(await ds.MetricFindQuery({ query: 'CollectionLookup(pd,prod-server,np,nonprod-server,dev,dev-server,np)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
-      expect(await ds.MetricFindQuery({ query: 'CollectionLookup(A,a,B,b,C,c,D,d,C)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
-      expect(await ds.MetricFindQuery({ query: 'CollectionLookup(A,a,B,b,C,c,D,d,E)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
+      expect(await ds.getVariableQueryValues({ query: 'CollectionLookup(pd,prod-server,np,nonprod-server,dev,dev-server,np)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
+      expect(await ds.getVariableQueryValues({ query: 'CollectionLookup(A,a,B,b,C,c,D,d,C)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
+      expect(await ds.getVariableQueryValues({ query: 'CollectionLookup(A,a,B,b,C,c,D,d,E)', queryType: 'legacy', refId: 'variable' })).toMatchSnapshot();
     });
   });
 });
